+++ argos/widgets/playingbox.py
@@ -6,6 +6,7 @@ from typing import List, Optional
 
 from gi.repository import GLib, GObject, Gtk
 
+from argos.download import ImageDownloader
 from argos.model import PlaybackState
 from argos.widgets.playingboxemptytracklistbox import PlayingBoxEmptyTracklistBox
 from argos.widgets.tracklengthbox import TrackLengthBox
@@ -68,6 +69,7 @@ class PlayingBox(Gtk.Box):
 
         self._app = application
         self._model = application.model
+        self._download: ImageDownloader = application.props.download
         self._disable_tooltips = application.props.disable_tooltips
 
         track_length_box = TrackLengthBox(application)
@@ -107,8 +109,9 @@ class PlayingBox(Gtk.Box):
         )
         self._model.playback.connect("notify::state", self._update_play_button)
         self._model.playback.connect(
-            "notify::image-path", self._update_playing_track_image
+            "notify::image-uri", self._reset_playing_track_image
         )
+        self._download.connect("image-downloaded", self._update_playing_track_image)
 
         self.show_all()
 
@@ -166,11 +169,24 @@ class PlayingBox(Gtk.Box):
             self._update_track_name_label()
             self._update_artist_name_label()
 
-    def _update_playing_track_image(
+    def _reset_playing_track_image(
         self,
         _1: GObject.GObject,
         _2: GObject.GParamSpec,
     ) -> None:
+        self.playing_track_image.set_from_pixbuf(self.default_track_image)
+        self.playing_track_image.show_now()
+
+    def _update_playing_track_image(
+        self,
+        _1: ImageDownloader,
+        image_uri: str,
+    ) -> None:
+        LOGGER.debug("Updating playing track image")
+        match_playing_track = self._model.playback.image_uri == image_uri
+        if not match_playing_track:
+            return
+
         image_path = (
             Path(self._model.playback.image_path)
             if self._model.playback.image_path